// ged-angular-www-ui/src/app/api/generated/model/book.ts
/**
 * WWW Angular OpenApi Demo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: dominik.mathmann@gedoplan.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {Author} from './author';


export interface Book {
  authors?: Array<Author>;
  id?: number;
  isbn?: string;
  pageCount?: number;
  publisher?: number;
  title?: string;
}